fix: value conversion error in app raffle

+++ AppOperadorEndpoints/Assets/Scripts/Settings/GameManager.cs
@@ -122,7 +122,7 @@ public class GameManager : MonoBehaviour
             yield return null;
         }
     }
-    public string FormatMoneyInfo(int value, int decimalHouse = 2)
+    public string FormatMoneyInfo(float value, int decimalHouse = 2)
     {
         string prizeFormated = string.Format(CultureInfo.CurrentCulture, value.ToString($"C{decimalHouse}"));
         return prizeFormated;
